+++ NEWS.md
@@ -1,3 +1,16 @@
+# profilePro 0.8.0
+
+* Added missing import of `MSnbase::as.MSnExp.OnDiskMSnExp()`.
+
+* Re-export [`plan()`](https://future.futureverse.org/reference/plan.html) from the [`future`](https://future.futureverse.org/) package.
+
+* The parallel processing back-end for xcms processing is now specified using [`plan()`](https://future.futureverse.org/reference/plan.html) from [`future`](https://future.futureverse.org/) package.
+This is enabled by the [`BiocParallel.FutureParam`](https://biocparallel.futureparam.futureverse.org/) package.
+
+* Removed the `nCores` argument from [`profileParameters()`](https://jasenfinch.github.io/profilePro/reference/profileParameters.html).
+
+* Removed [`patchwork`](https://patchwork.data-imaginist.com/) as a dependency.
+
 # profilePro 0.7.1
 
 * Removed parallel arguments from [`erah`](https://github.com/xdomingoal/erah-devel) processing routine as these now support the use of the [`future`](https://future.futureverse.org/) package.